+++ packages/caip/src/caip19/caip19.ts
@@ -3,7 +3,12 @@ import { ChainTypes, NetworkTypes } from '@shapeshiftoss/types'
 
 import { ChainNamespace, ChainReference, toCAIP2 } from '../caip2/caip2'
 
+/**
+ * @deprecated - Temporarily left in place for backwards compatibility, to be replaced with AssetId
+ */
 export type CAIP19 = string
+
+export type AssetId = string
 
 export enum AssetNamespace {
   CW20 = 'cw20',