+++ packages/swapper/src/swappers/zrx/ZrxSwapper.test.ts
@@ -1,6 +1,6 @@
 import { ChainAdapterManager } from '@shapeshiftoss/chain-adapters'
 import { HDWallet } from '@shapeshiftoss/hdwallet-core'
-import { chainAdapters, ChainTypes, GetQuoteInput, Quote, SwapperType } from '@shapeshiftoss/types'
+import { ChainTypes, GetQuoteInput, Quote, SwapperType } from '@shapeshiftoss/types'
 import Web3 from 'web3'
 
 import { ZrxError } from '../..'
@@ -8,9 +8,8 @@ import { ZrxSwapper } from '..'
 import { ZrxBuildQuoteTx } from '../zrx/ZrxBuildQuoteTx/ZrxBuildQuoteTx'
 import { getZrxMinMax } from './getZrxMinMax/getZrxMinMax'
 import { getZrxQuote } from './getZrxQuote/getZrxQuote'
-import { getZrxSendMaxAmount } from './getZrxSendMaxAmount/getZrxSendMaxAmount'
 import { getUsdRate } from './utils/helpers/helpers'
-import { BTC, FOX, WETH } from './utils/test-data/assets'
+import { FOX } from './utils/test-data/assets'
 import { setupQuote } from './utils/test-data/setupSwapQuote'
 import { ZrxApprovalNeeded } from './ZrxApprovalNeeded/ZrxApprovalNeeded'
 import { ZrxApproveInfinite } from './ZrxApproveInfinite/ZrxApproveInfinite'
@@ -41,13 +40,9 @@ jest.mock('./ZrxApproveInfinite/ZrxApproveInfinite', () => ({
   ZrxApproveInfinite: jest.fn()
 }))
 
-jest.mock('./getZrxSendMaxAmount/getZrxSendMaxAmount', () => ({
-  getZrxSendMaxAmount: jest.fn()
-}))
-
 describe('ZrxSwapper', () => {
   const input = <GetQuoteInput>{}
-  const quote = <Quote<ChainTypes, SwapperType>>{}
+  const quote = <Quote<ChainTypes>>{}
   const wallet = <HDWallet>{}
   const web3 = <Web3>{}
   const adapterManager = <ChainAdapterManager>{}
@@ -69,21 +64,6 @@ describe('ZrxSwapper', () => {
     const error = new ZrxError(message)
     expect(error.message).toBe(`ZrxError:${message}`)
   })
-  it('getAvailableAssets filters out all non-ethereum assets', () => {
-    const swapper = new ZrxSwapper(zrxSwapperDeps)
-    const availableAssets = swapper.getAvailableAssets([BTC, FOX, WETH])
-    expect(availableAssets).toStrictEqual([FOX, WETH])
-  })
-  it('canTradePair fails on non-eth chains', () => {
-    const swapper = new ZrxSwapper(zrxSwapperDeps)
-    const canTradePair = swapper.canTradePair(BTC, WETH)
-    expect(canTradePair).toBeFalsy()
-  })
-  it('canTradePair succeeds on eth chains', () => {
-    const swapper = new ZrxSwapper(zrxSwapperDeps)
-    const canTradePair = swapper.canTradePair(FOX, WETH)
-    expect(canTradePair).toBeTruthy()
-  })
   it('calls ZrxBuildQuoteTx on swapper.buildQuoteTx', async () => {
     const swapper = new ZrxSwapper(zrxSwapperDeps)
     const args = { input, wallet }
@@ -102,7 +82,7 @@ describe('ZrxSwapper', () => {
     const ethCAIP19 = 'eip155:1/slip44:60'
     expect(pair).toHaveLength(2)
     expect(pair[0]).toEqual(ethCAIP19)
-    expect(pair[1]).toEqual(FOX.caip19)
+    expect(pair[1]).toEqual(FOX.assetId)
   })
   it('calls getUsdRate on swapper.getUsdRate', async () => {
     const swapper = new ZrxSwapper(zrxSwapperDeps)
@@ -129,17 +109,5 @@ describe('ZrxSwapper', () => {
     const args = { quote, wallet }
     await swapper.approveInfinite(args)
     expect(ZrxApproveInfinite).toHaveBeenCalled()
-  })
-
-  it('calls getZrxSendMaxAmount on swapper.getSendMaxAmount', async () => {
-    const swapper = new ZrxSwapper(zrxSwapperDeps)
-    const args = {
-      quote,
-      wallet,
-      sellAssetAccountId: '0',
-      feeEstimateKey: chainAdapters.FeeDataKey.Average
-    }
-    await swapper.getSendMaxAmount(args)
-    expect(getZrxSendMaxAmount).toHaveBeenCalled()
   })
 })