+++ packages/swapper/src/swappers/zrx/ZrxSwapper.ts
@@ -12,20 +12,19 @@ import {
   GetQuoteInput,
   MinMaxOutput,
   Quote,
-  SendMaxAmountInput,
   SwapperType
 } from '@shapeshiftoss/types'
 import Web3 from 'web3'
 
-import { Swapper } from '../../api'
+import { BuyAssetBySellIdInput, Swapper } from '../../api'
 import { getZrxMinMax } from './getZrxMinMax/getZrxMinMax'
 import { getZrxQuote } from './getZrxQuote/getZrxQuote'
-import { getZrxSendMaxAmount } from './getZrxSendMaxAmount/getZrxSendMaxAmount'
 import { getUsdRate } from './utils/helpers/helpers'
 import { ZrxApprovalNeeded } from './ZrxApprovalNeeded/ZrxApprovalNeeded'
 import { ZrxApproveInfinite } from './ZrxApproveInfinite/ZrxApproveInfinite'
 import { ZrxBuildQuoteTx } from './ZrxBuildQuoteTx/ZrxBuildQuoteTx'
 import { ZrxExecuteQuote } from './ZrxExecuteQuote/ZrxExecuteQuote'
+
 export type ZrxSwapperDeps = {
   adapterManager: ChainAdapterManager
   web3: Web3
@@ -50,11 +49,11 @@ export class ZrxSwapper implements Swapper {
     return SwapperType.Zrx
   }
 
-  async buildQuoteTx(args: BuildQuoteTxInput): Promise<Quote<ChainTypes, SwapperType>> {
+  async buildQuoteTx(args: BuildQuoteTxInput): Promise<Quote<ChainTypes>> {
     return ZrxBuildQuoteTx(this.deps, args)
   }
 
-  async getQuote(input: GetQuoteInput): Promise<Quote<ChainTypes, SwapperType>> {
+  async getQuote(input: GetQuoteInput): Promise<Quote<ChainTypes>> {
     return getZrxQuote(input)
   }
 
@@ -66,36 +65,31 @@ export class ZrxSwapper implements Swapper {
     return getZrxMinMax(input)
   }
 
-  getAvailableAssets(assets: Asset[]): Asset[] {
-    return assets.filter((asset) => asset.chain === ChainTypes.Ethereum)
-  }
-
-  canTradePair(sellAsset: Asset, buyAsset: Asset): boolean {
-    const availableAssets = this.getAvailableAssets([sellAsset, buyAsset])
-    return availableAssets.length === 2
-  }
-
   getDefaultPair(): [CAIP19, CAIP19] {
     const ETH = 'eip155:1/slip44:60'
     const FOX = 'eip155:1/erc20:0xc770eefad204b5180df6a14ee197d99d808ee52d'
     return [ETH, FOX]
   }
 
-  async executeQuote(args: ExecQuoteInput<ChainTypes, SwapperType>): Promise<ExecQuoteOutput> {
+  async executeQuote(args: ExecQuoteInput<ChainTypes>): Promise<ExecQuoteOutput> {
     return ZrxExecuteQuote(this.deps, args)
   }
 
-  async approvalNeeded(
-    args: ApprovalNeededInput<ChainTypes, SwapperType>
-  ): Promise<ApprovalNeededOutput> {
+  async approvalNeeded(args: ApprovalNeededInput<ChainTypes>): Promise<ApprovalNeededOutput> {
     return ZrxApprovalNeeded(this.deps, args)
   }
 
-  async approveInfinite(args: ApproveInfiniteInput<ChainTypes, SwapperType>): Promise<string> {
+  async approveInfinite(args: ApproveInfiniteInput<ChainTypes>): Promise<string> {
     return ZrxApproveInfinite(this.deps, args)
   }
 
-  async getSendMaxAmount(args: SendMaxAmountInput): Promise<string> {
-    return getZrxSendMaxAmount(this.deps, args)
+  filterBuyAssetsBySellAssetId(args: BuyAssetBySellIdInput): CAIP19[] {
+    const { assetIds, sellAssetId } = args
+    // TODO: pending changes to caip lib, we may want to import caip2 value instead.
+    return assetIds.filter((id) => id.startsWith('eip155:1') && sellAssetId.startsWith('eip155:1'))
+  }
+
+  filterAssetIdsBySellable(assetIds: CAIP19[]): CAIP19[] {
+    return assetIds.filter((id) => id.startsWith('eip155:1'))
   }
 }