+++ packages/types/src/chain-adapters/cosmos.ts
@@ -12,7 +12,7 @@ export type Info = {
 export type Validator = {
   address: string
   moniker: string
-  tokens: string
+  tokens?: string
   apr: string
   commission: string
 }